+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - Mapped clang-format config file (.clang-format) to YAML syntax (@TruncatedDinosour)
 - log syntax: improved handling of escape characters in double quoted strings. See #2123 (@keith-hall)
+- Associate `/var/spool/mail/*` and `/var/mail/*` with the `Email` syntax. See #2156 (@cyqsimon)
 - Added cmd-help syntax to scope --help messages. See #2148 (@victor-gp)
 
 ## Themes